Delete redundant code from prev API
